requirements.txt
yt-dlp>=2024.0.0
openai-whisper>=20231117
demucs>=4.0.0
pydub>=0.25.1
numpy>=1.24.0
pyinstaller>=6.0.0
